requirements.txt
pytesseract
pandas
Pillow
ipython
